Reorganize, add new test configuration for actions.
