+++ src/test/java/cloud/filibuster/functional/java_latency/JUnitFilibusterTestWithLatencyInjectionExpectedFailure.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.functional.java.latency;
+package cloud.filibuster.functional.java_latency;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -6,7 +6,7 @@ import cloud.filibuster.exceptions.filibuster.FilibusterAllowedTimeExceededExcep
 import cloud.filibuster.functional.JUnitBaseTest;
 import cloud.filibuster.instrumentation.helpers.Networking;
 import cloud.filibuster.junit.FilibusterTest;
-import cloud.filibuster.junit.server.latency.Filibuster1000msLatencyProfile;
+import cloud.filibuster.junit.configuration.FilibusterLatencyOnlyAnalysisConfigurationFile;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.StatusRuntimeException;
@@ -18,22 +18,20 @@ import org.junit.jupiter.api.TestMethodOrder;
 import java.util.concurrent.TimeUnit;
 
 import static cloud.filibuster.junit.Assertions.assertPassesWithinMsOrThrowsUnderFault;
-import static cloud.filibuster.junit.Assertions.wasFaultInjected;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 
 /**
  * Test simple annotation usage.
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterTestWithLatencyProfile extends JUnitBaseTest {
+public class JUnitFilibusterTestWithLatencyInjectionExpectedFailure extends JUnitBaseTest {
     /**
      * Inject faults between Hello and World using Filibuster and assert proper faults are injected.
      *
      * @throws InterruptedException if teardown of gRPC channel fails.
      */
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @FilibusterTest(latencyProfile=Filibuster1000msLatencyProfile.class, expected=FilibusterAllowedTimeExceededException.class)
+    @FilibusterTest(analysisConfigurationFile=FilibusterLatencyOnlyAnalysisConfigurationFile.class, expected=FilibusterAllowedTimeExceededException.class)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
@@ -41,12 +39,13 @@ public class JUnitFilibusterTestWithLatencyProfile extends JUnitBaseTest {
                 .usePlaintext()
                 .build();
 
-        assertPassesWithinMsOrThrowsUnderFault(1, StatusRuntimeException.class, () -> {
+        // 800ms+ without FI
+        // 1800ms+ with FI.
+        assertPassesWithinMsOrThrowsUnderFault(900, StatusRuntimeException.class, () -> {
             HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
             Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
             Hello.HelloReply reply = blockingStub.partialHello(request);
             assertEquals("Hello, Armerian World!!", reply.getMessage());
-            assertFalse(wasFaultInjected());
         });
 
         helloChannel.shutdownNow();
